requirements.txt
Django==1.11.4
django-allauth==0.33.0
django-autoslug==1.9.3
django-contrib-comments==1.8.0
django-robots==3.0
gunicorn==19.7.1
Pillow==4.2.1
psycopg2==2.7.3.1
whitenoise==3.3.0
celery==4.1.0
django-fluent-dashboard==0.6.1